pyproject.toml
[tool.poetry]
name = "invoice-etl"
version = "0.1.0"
description = "ETL pipeline for extracting transaction data from customer invoice PDFs into PostgreSQL."
authors = []
readme = "README.md"
packages = [{ include = "invoice_etl", from = "src" }]

[tool.poetry.scripts]
invoice-etl = "invoice_etl.main:main"

[tool.poetry.dependencies]
python = "^3.12"
pdfplumber = "^0.11"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
pydantic = "^2.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0"
ruff = "^0.4"
pyright = "^1.1"
pytest = "^8.0"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ---------------------------------------------------------------------------
# Black
# ---------------------------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py312"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["invoice_etl"]

# ---------------------------------------------------------------------------
# Pyright
# ---------------------------------------------------------------------------
[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
venvPath = "."
venv = ".venv"

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=invoice_etl --cov-report=term-missing"
